Add argument parsing to passgen
This commit allows user to use the password generator from the command
line. We also cleaned up some function parameters to use kwargs, as we
can't always be sure which functionalities are in use. This also should
prevent argument bloat and added functionalities will simply need new
implementations and checks if they are present.