Fix unterminated line issue
In the past as we wrote the password to a file, we would simply write
the contents of the password string. However doing so without adding a
newline character at the end would result in some shells (zsh and fish)
to show the file ending with a % character when file was read with `cat`.
This unexpected behaviour was due to how these shells handle file output
with unterminated lines.

More info can be found at
https://unix.stackexchange.com/questions/649739/why-am-i-seeing-as-the-line-end-when-running-cat-file-cmd